refactor: 쿠폰 관련 모듈을 배럴 파일로 정리하고 import 경로 수정
