+++ src/features/coupon/ui/CouponList/index.tsx
@@ -1,8 +1,7 @@
 import * as S from './CouponList.styled';
-import Description from '@/shared/ui/Description';
-import { ClientCouponType } from '@entities/coupon/type/coupon.type';
+import Description from '@shared/ui/Description';
+import { ClientCouponType, COUPON_RULE } from '@entities/coupon';
 import CouponItem from '@features/coupon/ui/CouponItem';
-import { COUPON_RULE } from '@/entities/coupon/constants/couponRule';
 
 interface CouponListProps {
   clientCoupons: ClientCouponType[];
